+++ src/CosmosDB/CosmosDB/ChangeLog.md
@@ -19,6 +19,8 @@
 -->
 
 ## Upcoming Release
+
+## Version 0.2.0
 Introduced support for throughput Migration, allowing custoers to migrate their resources from manually provisioned throughput to autoscaled throughput. Customers can use this feature using the following cmdlets:
                'Invoke-AzCosmosDBSqlContainerThroughputMigration', 
                'Invoke-AzCosmosDBSqlDatabaseThroughputMigration',